Move project to the Tuple folder
