+++ src/index.spec.js
@@ -123,7 +123,7 @@ const nmap = require('./Tuple/nmap')
 const map = require('./pointfree/map')
 const merge = require('./pointfree/merge')
 const option = require('./pointfree/option')
-const project = require('./pointfree/project')
+const project = require('./Tuple/project')
 const promap = require('./pointfree/promap')
 const race = require('./Async/race')
 const read = require('./Writer/read')
