+++ src/pointfree/index.spec.js
@@ -28,7 +28,7 @@ const nmap = require('../Tuple/nmap')
 const map = require('./map')
 const merge = require('../pointfree/merge')
 const option = require('./option')
-const project = require('./project')
+const project = require('../Tuple/project')
 const promap = require('./promap')
 const race = require('../Async/race')
 const read = require('../Writer/read')
